# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EduManage Mobile API
  description: |
    Toàn bộ các endpoint dành cho ứng dụng mobile (student, parent, staff).

    **Authentication:** Hầu hết endpoint yêu cầu JWT Bearer token.
    Lấy token qua `POST /api/mobile/auth/login` hoặc `POST /api/mobile/auth/zalo`.

    ```
    Authorization: Bearer <token>
    ```
  version: "1.0.0"
  contact:
    name: Web Team
servers:
  - url: /api/mobile
    description: Mobile API prefix
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Post:
      type: object
      properties:
        id: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        authorName: { type: string }
        authorRole: { type: string }
        category:
          type: string
          enum: [thong-bao, su-kien, hoat-dong, hoc-thuat]
        content: { type: string }
        imageUrl: { type: string, nullable: true }
        imageUrls:
          type: array
          items: { type: string }
        isPinned: { type: boolean }
        locationId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        reactions: { type: object, additionalProperties: { type: integer } }
        myReaction: { type: string, nullable: true }

    # ─── Notification ────────────────────────────────────────
...
security:
  - bearerAuth: []
paths:
...
  /news-feed:
    get:
      tags: [News Feed]
      summary: Danh sách bài đăng
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [thong-bao, su-kien, hoat-dong, hoc-thuat, all]
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  permissions: { type: object }
                  posts: { type: array, items: { $ref: "#/components/schemas/Post" } }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
    post:
      tags: [News Feed]
      summary: Tạo bài đăng mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, content]
              properties:
                category:
                  type: string
                  enum: [thong-bao, su-kien, hoat-dong, hoc-thuat]
                content: { type: string }
                imageUrl: { type: string, nullable: true }
                imageUrls: { type: array, items: { type: string } }
                locationId: { type: string, format: uuid, nullable: true }
      responses:
        "201":
          description: Bài đăng đã tạo
          content:
            application/json:
              schema:
                type: object
                properties:
                  post: { $ref: "#/components/schemas/Post" }
        "403":
          description: Không có quyền đăng bài
...